feat: add InteractsWithEmailAuthentication trait to manage email authentication
